[Oztechan/CCC#2597] Move iOS WatchersView to WatchersRootView
